🚸 Remove `interactions` add-on.
Some stories needed a disabled actions tab and when doing that, the whole page broke. This happened because the interactions add-on is a beta feature with some bugs that required the actions tab active all the time.

To prevent this problem, I decided to remove it from our installation until the tool functions correctly.